+++ ChangeLog.md
@@ -1,6 +1,7 @@
 # 1.1
 
- - Fixed Bootstrap Methods not visible during write ([#2](https://gitlab.com/fee1-dead/coffer/-/issues/2))
+ - Fixed Bootstrap Methods not visible during write 
  - Renamed `Error::MUTF` to `Error::Mutf`.
  - Added *Annotations variants to ClassAttribute
+ - `Dynamic` now doesn't `Clone` the `Rc` pointer and clones the inner value instead ([#2](https://gitlab.com/fee1-dead/coffer/-/issues/2))